Refactors how Function Metadata is handled in Mixer evaluator. (#1768)
Automatic merge from submit-queue.

Refactors how Function Metadata is handled in Mixer evaluator.

**What this PR does / why we need it**:
This PR refactors the internal function metadata model of the Mixer evaluator, to unblock upcoming feature changes.
Particularly:

Removes EvalString method from the Evaluator interface: This is simply not used at this point.
Changes the FuncBase interface to be a FunctionMetadata struct (i.e. a pure-data only struct that keeps the metadata).
Removes the internal dependency of the compiler to the extern function set: All the metadata for externs need to be supplied externally now.
Moves the TypeChecker instantiation code to the il/evaluator package, reducing the surface area, and also making sure that the metadata is supplied consistently to both TypeChecker and the Evaluator.

**Which issue this PR fixes** *(optional, in `fixes #<issue number>(, fixes #<issue_number>, ...)` format, will close that issue when PR gets merged)*: fixes #

**Special notes for your reviewer**:

**Release note**:

```release-note
NONE
```