[pilot] Add yaml file monitor (#1717)
Automatic merge from submit-queue.

[pilot] Add yaml file monitor

**What this PR does / why we need it**:
This adds the a monitor of a directory containing yaml files and updates
a ConfigStore when changes are detected.  This is modeled after mixer's
configstore2.

Also removing the previous file-based config store as the capabilities
of this monitor are a superset of those in the config store.

**Which issue this PR fixes**: fixes #1662

**Special notes for your reviewer**:

**Release note**:
```release-note
None
```