Add comprehensive function support to the IL code generator.  (#1811)
Automatic merge from submit-queue.

Add comprehensive function support to the IL code generator. 

**What this PR does / why we need it**:
The current implementation of IL code generator hard-wires function names and signatures. This PR makes the generator generic, and makes it use the type-information for generating IL.

Also, it fixes a code-generator bug for the "|" operator, where the LHS is a constant. (The bug is, the constant would be evaluated, but the generated call would still fall-through to the next expression in line).

**Which issue this PR fixes** *(optional, in `fixes #<issue number>(, fixes #<issue_number>, ...)` format, will close that issue when PR gets merged)*: fixes #

**Special notes for your reviewer**:

**Release note**:

```release-note
NONE
```