removed omitempty from Route TimeoutMS (#1748)
Automatic merge from submit-queue.

removed omitempty from Route TimeoutMS

the `omitempty` keyword in the Route definition is preventing an empty timeout (=0) to be defined. This leads to the Envoy Default Timeout of 15s which breakes the long lasting connections, like the ones used in gRPC Streaming.

With this change, the Timeout is alway 0 (no timeout) except when you set it to another value in a RouteRule.

**What this PR does / why we need it**:

**Which issue this PR fixes** *(optional, in `fixes #<issue number>(, fixes #<issue_number>, ...)` format, will close that issue when PR gets merged)*: fixes #

**Special notes for your reviewer**:

**Release note**:

```release-note
NONE
```