Avoid using StringTable for runtime data. (#1974)
Automatic merge from submit-queue.

Avoid using StringTable for runtime data.

**What this PR does / why we need it**:
This fixes an issue in the evaluator that causes runtime strings to be captured in the strings table. With this fix:
- The strings are stored in heap, similar to other interface{} objects.
- The comparisons are performed using regular string comparison.
- The size-based expression cache eviction algorithm (& its parameter) is removed.
- Renamed the method on StringTable from GetID => Add for clarity.

**Which issue this PR fixes** *(optional, in `fixes #<issue number>(, fixes #<issue_number>, ...)` format, will close that issue when PR gets merged)*: fixes #
fixes #1369
**Special notes for your reviewer**:
This will most likely conflict with the pkg/server change. I'll wait until that one lands, before checking this in.

**Release note**:

```release-note NONE
```