Introduce the mixer/pkg/server package. (#1911)
This new package absorbs the functionality previously implemented in
cmd/server. This package is what pulls all of the Mixer bits and pieces
together to form a running server.

This package is not only used for mainline Mixer, but can be used from
tests to create a "mixer in a box". As such, we no longer need the
mixer/pkg/mock package.