Fix pkg/log timestamp generation logic (#2321)
Automatic merge from submit-queue.

Fix pkg/log timestamp generation logic

The casting to `byte` was problematic for several of the values in the calculation. Anything that could be over 256 would cause incorrect results. This PR updates to cast after the math has been completed. It adds a few small tests to check year and microseconds (which were the only affected results).