Add function parameter support to mixer/pkg/pool.GoRoutinePool (#2376)
Automatic merge from submit-queue.

Add function parameter support to mixer/pkg/pool.GoRoutinePool

The current model of scheduling func(){} types causes the caller to
(almost) always create closure over the state that the function needs
to operate on.

This PR introduces the ability to schedule functions that takes
parameters and also the ability to pass those parameters on the side.